Automatic detect wireless interface.
The server (Orange Pi Zero in this case) just has a single Wi-Fi
adapter, so we can do an automatic detection and bind directly the
interface. But for the client, we still need to improve this feature
because the user can have more WiFi adapater in their PC/laptop.
Maybe implement a popover in GUI to select a right wireless interface.